dsa-indexer(metadata-move,mechanical_provable): extract indexer metadata
